Fix UI and UX
Made it so you can add an image as a document to your logs.

Made is so you can scroll on the sheet when you are adding a new log

Cleaned up spelling mistakes on the How To steps

Cleaned up words on so that people could understand how to add a vehicle or log if they have never added one before.